updated components and added description in readme
