+++ README.md
@@ -1,1 +1,9 @@
 # ag-grid-features
+
+## TableWithColumnSelector
+
+Simplified version of the [AG Grid Tool Panel](https://www.ag-grid.com/react-data-grid/component-tool-panel/) for column selection.
+
+## TableWithColumnSelector2
+
+Simplified version of the [AG Grid Tool Panel](https://www.ag-grid.com/react-data-grid/component-tool-panel/) for column selection. This components [lifts the state up](https://react.dev/learn/sharing-state-between-components) from the `SearchableList` component so it can work properly. This component was created for comparison purposes.
