+++ src/components/TableWithColumnSelector/SearchableList.jsx
@@ -1,17 +1,8 @@
-import { useState } from "react";
+import { useState, useRef } from "react";
 import PropTypes from "prop-types";
 import DragIcon from "../../icons/DragIcon";
 
-export default function SearchableList({
-  list,
-  toggleItem,
-  sortByCheck,
-  sortByName,
-  isDragging,
-  dragEnter,
-  dragStart,
-  dropItem,
-}) {
+export default function SearchableList({ list, setList }) {
   const [searchInput, setSearchInput] = useState("");
   const renderList =
     searchInput !== ""
@@ -19,6 +10,45 @@ export default function SearchableList({
           item.name.toLowerCase().includes(searchInput.toLowerCase())
         )
       : list;
+
+  const handleToggle = (name) => {
+    const newList = list.map((row) =>
+      row.name === name ? { ...row, checked: !row.checked } : row
+    );
+    setList(newList);
+  };
+
+  const sortByCheck = () => {
+    const newList = [...list];
+    newList.sort((a, b) => b.checked - a.checked);
+    setList(newList);
+  };
+
+  const sortByName = () => {
+    const newList = [...list];
+    newList.sort((a, b) => a.name.localeCompare(b.name));
+    setList(newList);
+  };
+
+  const draggingItem = useRef();
+  const dragOverItem = useRef();
+  const [isDragging, setIsDragging] = useState(false);
+
+  const handleDragStart = (event, position) => {
+    draggingItem.current = position;
+    setIsDragging(true);
+  };
+
+  const handleDragEnter = (event, position) => {
+    dragOverItem.current = position;
+    const listCopy = [...list];
+    const draggingItemContent = listCopy[draggingItem.current];
+    listCopy.splice(draggingItem.current, 1);
+    listCopy.splice(dragOverItem.current, 0, draggingItemContent);
+    draggingItem.current = dragOverItem.current;
+    dragOverItem.current = null;
+    setList(listCopy);
+  };
 
   return (
     <>
@@ -46,17 +76,17 @@ export default function SearchableList({
             <li
               key={itemId}
               draggable={true}
-              onDragStart={(event) => dragStart(event, index)}
-              onDragEnter={(event) => dragEnter(event, index)}
+              onDragStart={(event) => handleDragStart(event, index)}
+              onDragEnter={(event) => handleDragEnter(event, index)}
               onDragOver={(event) => event.preventDefault()}
-              onDrop={dropItem}
+              onDrop={() => setIsDragging(false)}
               className={isDragging ? "dragging" : null}
             >
               <input
                 type="checkbox"
                 id={itemId}
                 checked={checked}
-                onChange={() => toggleItem(name)}
+                onChange={() => handleToggle(name)}
               />
               <DragIcon className="icon" />
               <label htmlFor={itemId}>{name}</label>
@@ -71,15 +101,9 @@ export default function SearchableList({
 SearchableList.propTypes = {
   list: PropTypes.arrayOf(
     PropTypes.shape({
-      checked: PropTypes.bool.isRequired,
-      name: PropTypes.string.isRequired,
+      checked: PropTypes.bool,
+      name: PropTypes.string,
     })
   ),
-  toggleItem: PropTypes.func.isRequired,
-  sortByCheck: PropTypes.func.isRequired,
-  sortByName: PropTypes.func.isRequired,
-  isDragging: PropTypes.bool.isRequired,
-  dragEnter: PropTypes.func.isRequired,
-  dragStart: PropTypes.func.isRequired,
-  dropItem: PropTypes.func.isRequired,
+  setList: PropTypes.func,
 };
