+++ src/components/TableWithColumnSelector2/SearchableList.jsx
@@ -1,9 +1,17 @@
-import { useState, useRef } from "react";
+import { useState } from "react";
 import PropTypes from "prop-types";
 import DragIcon from "../../icons/DragIcon";
 
-export default function SearchableList({ defaultList }) {
-  const [list, setList] = useState(defaultList);
+export default function SearchableList({
+  list,
+  toggleItem,
+  sortByCheck,
+  sortByName,
+  isDragging,
+  dragEnter,
+  dragStart,
+  dropItem,
+}) {
   const [searchInput, setSearchInput] = useState("");
   const renderList =
     searchInput !== ""
@@ -11,45 +19,6 @@ export default function SearchableList({ defaultList }) {
           item.name.toLowerCase().includes(searchInput.toLowerCase())
         )
       : list;
-
-  const handleToggle = (name) => {
-    const newList = list.map((row) =>
-      row.name === name ? { ...row, checked: !row.checked } : row
-    );
-    setList(newList);
-  };
-
-  const sortByCheck = () => {
-    const newList = [...list];
-    newList.sort((a, b) => b.checked - a.checked);
-    setList(newList);
-  };
-
-  const sortByName = () => {
-    const newList = [...list];
-    newList.sort((a, b) => a.name.localeCompare(b.name));
-    setList(newList);
-  };
-
-  const draggingItem = useRef();
-  const dragOverItem = useRef();
-  const [isDragging, setIsDragging] = useState(false);
-
-  const handleDragStart = (event, position) => {
-    draggingItem.current = position;
-    setIsDragging(true);
-  };
-
-  const handleDragEnter = (event, position) => {
-    dragOverItem.current = position;
-    const listCopy = [...list];
-    const draggingItemContent = listCopy[draggingItem.current];
-    listCopy.splice(draggingItem.current, 1);
-    listCopy.splice(dragOverItem.current, 0, draggingItemContent);
-    draggingItem.current = dragOverItem.current;
-    dragOverItem.current = null;
-    setList(listCopy);
-  };
 
   return (
     <>
@@ -77,17 +46,17 @@ export default function SearchableList({ defaultList }) {
             <li
               key={itemId}
               draggable={true}
-              onDragStart={(event) => handleDragStart(event, index)}
-              onDragEnter={(event) => handleDragEnter(event, index)}
+              onDragStart={(event) => dragStart(event, index)}
+              onDragEnter={(event) => dragEnter(event, index)}
               onDragOver={(event) => event.preventDefault()}
-              onDrop={() => setIsDragging(false)}
+              onDrop={dropItem}
               className={isDragging ? "dragging" : null}
             >
               <input
                 type="checkbox"
                 id={itemId}
                 checked={checked}
-                onChange={() => handleToggle(name)}
+                onChange={() => toggleItem(name)}
               />
               <DragIcon className="icon" />
               <label htmlFor={itemId}>{name}</label>
@@ -100,10 +69,17 @@ export default function SearchableList({ defaultList }) {
 }
 
 SearchableList.propTypes = {
-  defaultList: PropTypes.arrayOf(
+  list: PropTypes.arrayOf(
     PropTypes.shape({
-      checked: PropTypes.bool,
-      name: PropTypes.string,
+      checked: PropTypes.bool.isRequired,
+      name: PropTypes.string.isRequired,
     })
   ),
+  toggleItem: PropTypes.func.isRequired,
+  sortByCheck: PropTypes.func.isRequired,
+  sortByName: PropTypes.func.isRequired,
+  isDragging: PropTypes.bool.isRequired,
+  dragEnter: PropTypes.func.isRequired,
+  dragStart: PropTypes.func.isRequired,
+  dropItem: PropTypes.func.isRequired,
 };
